Google.php - multiple updates
The google provider used a deprecated scope, did not make use of the verified email and imported mail contacts only

google recommends the following scopes :
https://developers.google.com/+/api/oauth#scopes

It is a bit hard to define a google provider because google consists of so many different services and APIs. However:
hybridauth should at least follow the new OAuth scopes to support google+  but leave it to the user to customize the scopes ...

1. Updated scopes and made them dynamic
- auth/plus.login as the main log in scope
- auth/userinfo.email for verified email (optional now)
- /m8/feeds/ for gmail contacts (optional now)

[if a user does not want to use contacts, and just wants a plain profile, he can use .../auth/plus.me as only scope as well,
see http://hybridauth.sourceforge.net/support.html?place=msg%2Fhybridauth%2F9RL9irlM7ws%2FLz87NyLq9ZoJ
]

The advantage of the new login scope is that the user can select which social connections hybridauth is able to use (e.g. which contacts we get)
wrapped in a brilliant UI.

2. Makes use of verified mail now depending on the scope 
- by making a request to /userinfo

3. Supports google+ now
- google+ profile
- google+ contacts

4. Updated API versions 
- v1 for "new scopes"/google+
- v2 for userinfo
- v3 for gmail contacts