Enable extra parameters from config to be passed into OAuth1 authorizeUrl.
It is often necessary to request the permissions of a given service during authorization, and the OAuth1Client allows for an "extras" parameter, so I am passing the provider's config["extras"] array to the api's authorizeUrl method. This may also benefit the OAuth2 provider, but I haven't gone there yet.