Create Tencent.php
This module integration Tencent QQ OAuth2 login feature in hybridauth. 
As of 20 March 2013, there are 798.2 million active QQ zccounts, with a peek of 176.4 million simultaneous onlie QQ users. Most of these users come from China, and they are eager to learn and exchange with you.
This is my first work in git, I hope you enjoy it.