Fix properties and values methods in Collection
Collection::properties() returned property values instead of property
names as it says in the method's documentation. This resulted in the
Slack provider to not being able to iterate over the property names.

The only other place where Collection::properties() was used was in the
TwitchTV provider. This is updated to use the Collection::values() method
instead which returns the property values as expected.