Clear invalid authorization_state during OAuth2 login
There are some rare interactions where the cached OAuth2 state can get used without the login flow succeeding, like if a timeout happens or the user hits back at just the wrong moment, or other cases which I do not understand. If this happens, the browser gets jammed and cannot login, because it's stuck with an old session cookie with the invalid state.

Erase that invalid state when it is discovered.

This is part of addressing https://github.com/hybridauth/hybridauth/issues/1301. This isn't a full fix because while this defends against the app being jammed, I do not know what the real triggers are and if there is a more direct way to avoid this.